requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
Django==2.2
django-elasticsearch-dsl==7.2.2
django-environ==0.9.0
elasticsearch==7.17.6
elasticsearch-dsl==7.4.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.5
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
